refactor faculty schedule
