+++ client-course-schedulizer/src/components/Tabs/FacultySchedule/index.ts
@@ -1,0 +1,1 @@
+export * from "./FacultySchedule";
